refactor dag_visualizer example
